# apps/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KduFoot API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/debug-db": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Database Connectivity Debugging",
        "description": "Verifies the connection to the Cloudflare D1 database by performing a simple count query on the users table. Returns the database status and environment details. Warning: This endpoint should be restricted or disabled in production.\n",
        "responses": {
          "200": {
            "description": "Database connection is successful."
          },
          "500": {
            "description": "Database connection failed or query error."
          }
        }
      }
    },
...
    "/api/test-push": {
      "get": {
        "tags": [
          "Development"
        ],
        "summary": "Test Push Notification",
        "description": "Sends a dummy push notification to the current user to verify connectivity and VAPID setup.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Push attempt initiated."
          }
        }
      }
    },
    "/api/get_users": {
      "get": {
        "tags": [
          "Development"
        ],
        "summary": "Legacy Current User Identification",
        "description": "Returns the Auth0 'sub' claim for the currently authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user mapping identification."
          }
        }
      }
    },
    "/api/get/{user}": {
      "get": {
        "tags": [
          "Development"
        ],
        "summary": "Parameterized Identity Test",
        "description": "A test endpoint that accepts a user identifier in the path and returns it along with the current token and permissions. Used for verifying path parameter parsing and JWT context availability.\n",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "description": "Identifiant de l'utilisateur à tester.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Echo of provided parameters and authenticating token details."
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}